Added propTypes to MenuBar component
Separated sub components to separate files
